+++ facade/js/pages/homepage.js
@@ -1,5 +1,10 @@
 import { log } from "core";
+import $ from "jquery";
+import slideshow from "./../components/slideshow";
 
 export default function () {
     log("Homepage");
+
+    // setup Slick
+    slideshow($("header .slideshow"));
 }